Fix unterminated string in BadPattern exception
Previously, we would overallocate the string passed back to the OCaml
runtime containing the PCRE2 error message; now, we copy the raw error
to an intermediate buffer on the stack before allocating an OCaml string
of exactly the proper length. This change both fixes the absence of the
string terminator and consumes less memory overall.

Additionally, I've increased the size of the error message buffer to
256; while 128 is sufficient for the current set of errors supported by
PCRE2, the maximum length of any of which is 92 including the NUL
terminator, 256 gives us some breathing room should longer messages
begin to be used. This length is also a form of unofficial standard for
PCRE error message buffer sizes, being used in the Julia runtime and
other high-profile open-source applications. Since the OCaml string we
allocate will be exactly the length of the actual error message, the
impact this additional stack allocation has is negligible.